update js to ts
