+++ stripe_billing_dashboard/tasks/refund_stripe_charge.task.yaml
@@ -12,7 +12,7 @@ parameters:
   description: The Stripe ID for the charge to refund
   required: false
 node:
-  entrypoint: refund_stripe_charge.js
+  entrypoint: refund_stripe_charge.ts
   nodeVersion: "18"
   # Uncomment if you want to use your own Stripe API key from a config variable
   # Get your Stripe API keys here: https://stripe.com/docs/keys#obtain-api-keys
