+++ stripe_billing_dashboard/tasks/refund_stripe_charge.ts
@@ -1,8 +1,7 @@
-import Stripe from "stripe";
 // Remove the test Stripe API key below and replace with your own from a config variable
 // How to use config variables in Airplane: https://docs.airplane.dev/platform/configs
 const stripeAPIKey = process.env.STRIPE_SECRET_KEY ?? 'sk_test_4eC39HqLyjWDarjtT1zdp7dc';
-const stripe = Stripe(stripeAPIKey);
+const stripe = require('stripe')(stripeAPIKey);
 
 export default async function(params) {
   const refund = await stripe.refunds.create({
